requirements.txt
flask
python-dotenv
flask-restful
flask_sqlalchemy
flask_jwt_extended
flask_mail